apps/data-processor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.2
torch==2.1.0
transformers==4.34.1
python-dotenv==1.0.0
pytest==7.4.3
azure-cosmos==4.5.1 